+++ src/init/consented/messenger.ts
@@ -1,4 +1,4 @@
-import { initMessenger } from 'core';
+import { init as initMessenger } from 'core/messenger';
 import { init as background } from 'core/messenger/background';
 import { init as disableRefresh } from 'core/messenger/disable-refresh';
 import { init as fullwidth } from 'core/messenger/full-width';
@@ -12,7 +12,6 @@ import { init as resize } from 'core/messenger/resize';
 import { init as scroll } from 'core/messenger/scroll';
 import { init as type } from 'core/messenger/type';
 import { init as viewport } from 'core/messenger/viewport';
-import { reportError } from 'utils/report-error';
 
 /**
  * Messenger gets to skip the promise chain and run immediately.
@@ -32,7 +31,6 @@ initMessenger(
 		passback,
 	],
 	[scroll, viewport],
-	reportError,
 );
 
 const init = () => Promise.resolve();