+++ src/init/consentless.ts
@@ -1,5 +1,5 @@
 import type { ConsentState } from '@guardian/libs';
-import { initMessenger } from 'core';
+import { init as initMessenger } from 'core/messenger';
 import { init as background } from 'core/messenger/background';
 import { init as resize } from 'core/messenger/resize';
 import { init as type } from 'core/messenger/type';
@@ -10,11 +10,10 @@ import { initConsentless } from 'init/consentless/prepare-ootag';
 import { reloadPageOnConsentChange } from 'init/shared/reload-page-on-consent-change';
 import { init as setAdTestCookie } from 'init/shared/set-adtest-cookie';
 import { init as setAdTestInLabelsCookie } from 'init/shared/set-adtest-in-labels-cookie';
-import { reportError } from 'utils/report-error';
 
 const bootConsentless = async (consentState: ConsentState): Promise<void> => {
 	const consentlessModuleList = [
-		initMessenger([background, resize, type], [], reportError),
+		initMessenger([background, resize, type], []),
 		setAdTestCookie(),
 		setAdTestInLabelsCookie(),
 		initConsentless(consentState),