+++ src/insert/fill-dynamic-advert-slot.ts
@@ -1,12 +1,12 @@
 import { log } from '@guardian/libs';
 import type { SizeMapping } from 'core/ad-sizes';
-import { reportError } from 'utils/report-error';
 import type { Advert } from '../define/Advert';
 import { createAdvert } from '../define/create-advert';
 import { enableLazyLoad } from '../display/lazy-load';
 import { loadAdvert } from '../display/load-advert';
 import { dfpEnv } from '../lib/dfp/dfp-env';
 import { queueAdvert } from '../lib/dfp/queue-advert';
+import { reportError } from '../utils/report-error';
 
 const displayAd = (advert: Advert, forceDisplay: boolean) => {
 	if (dfpEnv.shouldLazyLoad() && !forceDisplay) {
@@ -29,14 +29,9 @@ const fillDynamicAdSlot = (
 			if (dfpEnv.adverts.has(adSlot.id)) {
 				const errorMessage = `Attempting to add slot with exisiting id ${adSlot.id}`;
 				log('commercial', errorMessage);
-				reportError(
-					Error(errorMessage),
-					{
-						feature: 'commercial',
-						slotId: adSlot.id,
-					},
-					false,
-				);
+				reportError(Error(errorMessage), 'commercial', {
+					slotId: adSlot.id,
+				});
 
 				return;
 			}