+++ src/utils/robust.ts
@@ -1,54 +1,31 @@
-/*
-    Swallows (and reports) exceptions. Designed to wrap around modules at the "bootstrap" level.
-    For example "comments throwing an exception should not stop auto refresh"
+/**
+ * Swallows and reports exceptions. Designed to wrap around modules at the "bootstrap" level.
  */
-
-import { convertError, reportError } from './report-error';
+import { reportError } from './report-error';
 
 type ModuleFunction = () => void;
 type Module = [string, ModuleFunction];
 type Modules = Module[];
 
-const catchErrors = (fn: ModuleFunction): Error | undefined => {
-	let error: Error | undefined;
-
-	try {
-		fn();
-	} catch (e) {
-		error = convertError(e);
-	}
-
-	return error;
-};
-
-const logError = (
-	moduleName: string,
-	error: Error,
-	tags?: Record<string, string>,
-): void => {
-	window.console.warn('Caught error.', error.stack);
-	reportError(error, { module: moduleName, ...tags }, false);
-};
-
-const catchAndLogError = (
-	name: string,
-	fn: ModuleFunction,
-	tags?: Record<string, string>,
-): void => {
-	const error = catchErrors(fn);
-
-	if (error) {
-		logError(name, error, tags);
-	}
-};
-
-const catchErrorsWithContext = (
+const catchErrorsAndReport = (
 	modules: Modules,
 	tags?: Record<string, string>,
 ): void => {
-	modules.forEach(([name, fn]) => catchAndLogError(name, fn, tags));
+	modules.forEach(([name, fn]) => {
+		let error: Error | undefined;
+
+		try {
+			fn();
+		} catch (e) {
+			error = e instanceof Error ? e : new Error(String(e));
+		}
+
+		if (error) {
+			window.console.warn('Caught error.', error.stack);
+			reportError(error, 'commercial', { ...tags, module: name });
+		}
+	});
 };
 
-export { catchErrorsWithContext };
+export { catchErrorsAndReport };
 export type { Modules };
-export const _ = { catchAndLogError };